Removed code for Salvage and Storage buttons.
Added barcode_2 button.

Fixed Return event bug. When return was pressed an event argument would
be sent to the callback function but the callback function didn't take
inputs. Fixed this by adding "lambda: event" so that the lambda
expression would take the event input and not input it into the
callback.

Added a Map event for the window. This allows the combobox to be
populated after the window has finished loading. A Map event calls the
callback once the window it's bound to has finished loading.

Added a single function that determines the input type of the button,
e.g., barcode 1, barcode 2, asset tag, or serial number.

Added options_scene().

Added populate_combobox() which calls the controller's
populate_combobox().