Added two jamf binary location vars and removed the hardcoded values from enroll(). These two vars can be set in the JSS server config.
Updated push_enroll_fields() to be more general so that it can work
with config files. The non-general aspect of it is that it gets the
previous name field which is an extension attribute of ours.

Added single line output to terminal to view enrolling process in
enroll() so that the output doesn't clutter the console but you're still
able to see progress.