Implementation specific to the user has been moved into user_actions.py.
To faciliate modifying the auto generated document, _build_data() was
developed. _build_data() takes in a list of tuples and generates most
of the body of the HTML file. user_actions.modify_items() is placed
right after it so that the user can further modify the list of tuples.
In the public version, modify_items() won't have an implementation, i.e.,
the body of the function will be "pass".